Align type with the one used in Maven plugin

+++ subprojects/jacoco/src/main/java/org/gradle/testing/jacoco/tasks/rules/JacocoLimit.java
@@ -21,6 +21,7 @@ import org.gradle.api.tasks.Input;
 import org.gradle.api.tasks.Optional;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Defines a Jacoco rule limit.
@@ -55,16 +56,16 @@ public interface JacocoLimit extends Serializable {
      */
     @Input
     @Optional
-    Double getMinimum();
+    BigDecimal getMinimum();
 
-    void setMinimum(Double minimum);
+    void setMinimum(BigDecimal minimum);
 
     /**
      * Gets the maximum expected value for limit. Default to null.
      */
     @Input
     @Optional
-    Double getMaximum();
+    BigDecimal getMaximum();
 
-    void setMaximum(Double maximum);
+    void setMaximum(BigDecimal maximum);
 }
